PAY-3573 : Liberata Pull APIs are unhooked from Internal Payment Swagger
