+++ api/src/main/java/uk/gov/hmcts/payment/api/controllers/liberata/LiberataPullApiController.java
@@ -1,0 +1,236 @@
+package uk.gov.hmcts.payment.api.controllers.liberata;
+
+import io.swagger.annotations.*;
+import org.ff4j.FF4j;
+import org.joda.time.LocalDateTime;
+import org.joda.time.format.DateTimeFormatter;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.bind.annotation.*;
+import uk.gov.hmcts.payment.api.configuration.LaunchDarklyFeatureToggler;
+import uk.gov.hmcts.payment.api.contract.PaymentDto;
+import uk.gov.hmcts.payment.api.contract.PaymentsResponse;
+import uk.gov.hmcts.payment.api.contract.util.Service;
+import uk.gov.hmcts.payment.api.controllers.PaymentExternalAPI;
+import uk.gov.hmcts.payment.api.dto.PaymentSearchCriteria;
+import uk.gov.hmcts.payment.api.dto.mapper.PaymentDtoMapper;
+import uk.gov.hmcts.payment.api.model.*;
+import uk.gov.hmcts.payment.api.service.CallbackService;
+import uk.gov.hmcts.payment.api.service.PaymentService;
+import uk.gov.hmcts.payment.api.util.DateUtil;
+import uk.gov.hmcts.payment.api.util.PaymentMethodType;
+import uk.gov.hmcts.payment.api.v1.model.exceptions.PaymentException;
+import uk.gov.hmcts.payment.api.v1.model.exceptions.PaymentNotFoundException;
+import uk.gov.hmcts.payment.api.validators.PaymentValidator;
+
+import java.math.BigDecimal;
+import java.util.*;
+import java.util.stream.Collectors;
+
+@RestController
+@Api(tags = {"Payment"})
+@SwaggerDefinition(tags = {@Tag(name = "LiberataPullApiController", description = "Payment REST API")})
+public class LiberataPullApiController {
+
+    private static final Logger LOG = LoggerFactory.getLogger(LiberataPullApiController.class);
+    private final PaymentService<PaymentFeeLink, String> paymentService;
+    private final PaymentDtoMapper paymentDtoMapper;
+    private final PaymentValidator validator;
+    private final FF4j ff4j;
+    private final DateTimeFormatter formatter;
+    private final PaymentFeeRepository paymentFeeRepository;
+    private final PaymentStatusRepository paymentStatusRepository;
+    private final CallbackService callbackService;
+
+    @Autowired
+    private LaunchDarklyFeatureToggler featureToggler;
+
+    @Autowired
+    public LiberataPullApiController(PaymentService<PaymentFeeLink, String> paymentService,
+                                        PaymentDtoMapper paymentDtoMapper,
+                                        PaymentValidator paymentValidator,
+                                        FF4j ff4j,
+                                        DateUtil dateUtil,
+                                        PaymentFeeRepository paymentFeeRepository,
+                                        PaymentStatusRepository paymentStatusRepository,
+                                        CallbackService callbackService) {
+        this.paymentService = paymentService;
+        this.paymentDtoMapper = paymentDtoMapper;
+        this.validator = paymentValidator;
+        this.ff4j = ff4j;
+        this.formatter = dateUtil.getIsoDateTimeFormatter();
+        this.paymentFeeRepository = paymentFeeRepository;
+        this.paymentStatusRepository = paymentStatusRepository;
+        this.callbackService = callbackService;
+    }
+
+
+    @ApiOperation(value = "Get payments for between dates", notes = "Get list of payments. You can optionally provide start date and end dates which can include times as well. Following are the supported date/time formats. These are yyyy-MM-dd, dd-MM-yyyy," +
+        "yyyy-MM-dd HH:mm:ss, dd-MM-yyyy HH:mm:ss, yyyy-MM-dd'T'HH:mm:ss, dd-MM-yyyy'T'HH:mm:ss")
+    @ApiResponses(value = {
+        @ApiResponse(code = 200, message = "Payments retrieved"),
+        @ApiResponse(code = 400, message = "Bad request")
+    })
+    @GetMapping(value = "/payments")
+    @PaymentExternalAPI
+    public PaymentsResponse retrievePayments(@RequestParam(name = "start_date", required = false) Optional<String> startDateTimeString,
+                                             @RequestParam(name = "end_date", required = false) Optional<String> endDateTimeString,
+                                             @RequestParam(name = "payment_method", required = false) Optional<String> paymentMethodType,
+                                             @RequestParam(name = "service_name", required = false) Optional<String> serviceType,
+                                             @RequestParam(name = "ccd_case_number", required = false) String ccdCaseNumber,
+                                             @RequestParam(name = "pba_number", required = false) String pbaNumber
+    ) {
+
+        if (!ff4j.check("payment-search")) {
+            throw new PaymentException("Payment search feature is not available for usage.");
+        }
+
+        validator.validate(paymentMethodType, serviceType, startDateTimeString, endDateTimeString);
+
+        Date fromDateTime = Optional.ofNullable(startDateTimeString.map(formatter::parseLocalDateTime).orElse(null))
+            .map(LocalDateTime::toDate)
+            .orElse(null);
+
+        Date toDateTime = Optional.ofNullable(endDateTimeString.map(formatter::parseLocalDateTime).orElse(null))
+            .map(s -> fromDateTime != null && s.getHourOfDay() == 0 ? s.plusDays(1).minusSeconds(1).toDate() : s.toDate())
+            .orElse(null);
+
+        List<PaymentFeeLink> paymentFeeLinks = paymentService
+            .search(
+                PaymentSearchCriteria
+                    .searchCriteriaWith()
+                    .startDate(fromDateTime)
+                    .endDate(toDateTime)
+                    .ccdCaseNumber(ccdCaseNumber)
+                    .pbaNumber(pbaNumber)
+                    .paymentMethod(paymentMethodType.map(value -> PaymentMethodType.valueOf(value.toUpperCase()).getType()).orElse(null))
+                    .serviceType(serviceType.map(value -> Service.valueOf(value.toUpperCase()).getName()).orElse(null))
+                    .build()
+            );
+
+        final List<PaymentDto> paymentDtos = new ArrayList<>();
+        LOG.info("No of paymentFeeLinks retrieved for Liberata Pull : {}", paymentFeeLinks.size());
+        for (final PaymentFeeLink paymentFeeLink: paymentFeeLinks) {
+            populatePaymentDtos(paymentDtos, paymentFeeLink);
+        }
+        return new PaymentsResponse(paymentDtos);
+    }
+
+    @ApiOperation(value = "Update payment status by payment reference", notes = "Update payment status by payment reference")
+    @ApiResponses(value = {
+        @ApiResponse(code = 204, message = "No content"),
+        @ApiResponse(code = 404, message = "Payment not found")
+    })
+    @PaymentExternalAPI
+    @PatchMapping("/payments/{reference}/status/{status}")
+    @Transactional
+    public ResponseEntity updatePaymentStatus(@PathVariable(value = "reference", required = true) String reference,
+                                              @PathVariable(value = "status", required = true) String status) {
+        Optional<Payment> payment = getPaymentByReference(reference);
+
+        if (payment.isPresent()) {
+            payment.get().setPaymentStatus(paymentStatusRepository.findByNameOrThrow(status));
+            if (payment.get().getServiceCallbackUrl() != null) {
+                callbackService.callback(payment.get().getPaymentLink(), payment.get());
+            }
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
+
+    private Optional<Payment> getPaymentByReference(String reference) {
+        PaymentFeeLink paymentFeeLink = paymentService.retrieve(reference);
+        return paymentFeeLink.getPayments().stream()
+            .filter(p -> p.getReference().equals(reference)).findAny();
+    }
+
+    private void populatePaymentDtos(final List<PaymentDto> paymentDtos, final PaymentFeeLink paymentFeeLink) {
+        //Adding this filter to exclude Exela payments if the bulk scan toggle feature is disabled.
+        List<Payment> payments = getFilteredListBasedOnBulkScanToggleFeature(paymentFeeLink);
+        boolean apportionFeature = featureToggler.getBooleanValue("apportion-feature",false);
+
+        LOG.info("BSP Feature ON : No of Payments retrieved for Liberata Pull : {}", payments.size());
+        LOG.info("Apportion feature flag in liberata API: {}", apportionFeature);
+        for (final Payment payment: payments) {
+            final String paymentReference = paymentFeeLink.getPaymentReference();
+            //Apportion logic added for pulling allocation amount
+            boolean apportionCheck = payment.getPaymentChannel() != null
+                && !payment.getPaymentChannel().getName().equalsIgnoreCase(Service.DIGITAL_BAR.getName());
+            LOG.info("Apportion check value in liberata API: {}", apportionCheck);
+            List<PaymentFee> fees = paymentFeeLink.getFees();
+            boolean isPaymentAfterApportionment = false;
+            if (apportionCheck && apportionFeature) {
+                LOG.info("Apportion check and feature passed");
+                final List<FeePayApportion> feePayApportionList = paymentService.findByPaymentId(payment.getId());
+                if(feePayApportionList != null && !feePayApportionList.isEmpty()) {
+                    LOG.info("Apportion details available in PaymentController");
+                    fees = new ArrayList<>();
+                    getApportionedDetails(fees, feePayApportionList);
+                    isPaymentAfterApportionment = true;
+                }
+            }
+            //End of Apportion logic
+            final PaymentDto paymentDto = paymentDtoMapper.toReconciliationResponseDtoForLibereta(payment, paymentReference, fees,ff4j,isPaymentAfterApportionment);
+            paymentDtos.add(paymentDto);
+        }
+    }
+
+    private void getApportionedDetails(List<PaymentFee> fees, List<FeePayApportion> feePayApportionList) {
+        LOG.info("Getting Apportionment Details!!!");
+        for (FeePayApportion feePayApportion : feePayApportionList)
+        {
+            Optional<PaymentFee> apportionedFee = paymentFeeRepository.findById(feePayApportion.getFeeId());
+            if(apportionedFee.isPresent())
+            {
+                LOG.info("Apportioned fee is present");
+                PaymentFee fee = apportionedFee.get();
+                if(feePayApportion.getApportionAmount() != null) {
+                    LOG.info("Apportioned Amount is available!!!");
+                    BigDecimal allocatedAmount = feePayApportion.getApportionAmount()
+                        .add(feePayApportion.getCallSurplusAmount() != null
+                            ? feePayApportion.getCallSurplusAmount()
+                            : BigDecimal.valueOf(0));
+                    LOG.info("Allocated amount in PaymentController: {}", allocatedAmount);
+                    fee.setAllocatedAmount(allocatedAmount);
+                    fee.setDateApportioned(feePayApportion.getDateCreated());
+                }
+                fees.add(fee);
+            }
+        }
+    }
+
+    private List<Payment> getFilteredListBasedOnBulkScanToggleFeature(PaymentFeeLink paymentFeeLink) {
+        List<Payment> payments = paymentFeeLink.getPayments();
+        boolean bulkScanCheck = ff4j.check("bulk-scan-check");
+        LOG.info("bulkScanCheck value: {}",bulkScanCheck);
+        if(!bulkScanCheck) {
+            LOG.info("BSP Feature OFF : No of Payments retrieved for Liberata Pull : {}", payments.size());
+            payments = Optional.ofNullable(payments)
+                .orElseGet(Collections::emptyList)
+                .stream()
+                .filter(payment -> Objects.nonNull(payment.getPaymentChannel()))
+                .filter(payment -> Objects.nonNull(payment.getPaymentChannel().getName()))
+                .filter(payment -> !payment.getPaymentChannel().getName().equalsIgnoreCase("bulk scan"))
+                .collect(Collectors.toList());
+        }
+        return payments;
+    }
+
+    @ResponseStatus(HttpStatus.NOT_FOUND)
+    @ExceptionHandler(PaymentNotFoundException.class)
+    public String notFound(PaymentNotFoundException ex) {
+        return ex.getMessage();
+    }
+
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    @ExceptionHandler(PaymentException.class)
+    public String return400(PaymentException ex) {
+        return ex.getMessage();
+    }
+
+}
